requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.10.0
click==8.1.7
Django==4.1.12
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
drf-spectacular==0.26.5
Faker==19.12.0
flake8==6.1.0
gunicorn==21.2.0
inflection==0.5.1
install==1.3.5
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
